Add files via upload
The dataset decision tree model is a machine learning algorithm that can be used to predict the class of an unknown data point. The model is trained on a dataset of known data points, and each data point is represented by a set of features. The model works by recursively partitioning the dataset into smaller subsets until each subset contains only data points with the same class label. The model can then be used to classify new data points by starting at the root of the tree and traversing the tree until a leaf node is reached.