Add files via upload
This dataset contains 14 attributes that describe various key indicators of heart disease, including age, sex, cholesterol level, and blood pressure. The goal is to predict whether or not a patient has heart disease, based on these indicators. The dataset was collected from 463 individuals, and the predictions were made using a random forest model. The model had an accuracy of 79.8%, meaning that it correctly predicted heart disease in 79.8% of the individuals in the dataset.